refactor(runtimed): split peer connection setup
