+++ crates/runtimed/src/notebook_sync_server/mod.rs
@@ -53,8 +53,9 @@ mod load;
 mod metadata;
 mod nbformat_convert;
 mod path_index;
-mod peer;
+mod peer_connection;
 mod peer_eviction;
+mod peer_loop;
 mod peer_notebook_sync;
 mod peer_pool_sync;
 mod peer_presence;
@@ -74,7 +75,7 @@ pub(crate) use load::*;
 pub(crate) use metadata::*;
 pub(crate) use nbformat_convert::*;
 pub use path_index::{PathIndex, PathIndexError};
-pub(crate) use peer::*;
+pub(crate) use peer_connection::*;
 #[cfg(test)]
 pub(crate) use peer_presence::sanitize_peer_label;
 pub(crate) use peer_runtime_agent::handle_runtime_agent_sync_connection;
